refactor all patient classes
