+++ src/test/java/seedu/address/logic/parser/patient/DeletePatientCommandParserTest.java
@@ -1,4 +1,4 @@
-package seedu.address.logic.parser;
+package seedu.address.logic.parser.patient;
 
 import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 import static seedu.address.logic.parser.CommandParserTestUtil.assertParseFailure;
@@ -7,7 +7,7 @@ import static seedu.address.testutil.TypicalIndexes.INDEX_FIRST_PERSON;
 
 import org.junit.jupiter.api.Test;
 
-import seedu.address.logic.commands.DeletePatientCommand;
+import seedu.address.logic.commands.patient.DeletePatientCommand;
 
 /**
  * As we are only doing white-box testing, our test cases do not cover path variations
